pyproject.toml
[project]
name = "recommendation-wayjet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.7.0",
    "fastapi>=0.100.0",
    "lightgbm>=3.3.0",
    "mlflow>=3.5.1",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "pyyaml>=6.0",
    "redis>=7.0.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "sentence-transformers>=2.2.0",
    "tqdm>=4.64.0",
    "uvicorn[standard]>=0.20.0",
]
